% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellmask}
\alias{cellmask}
\title{A thresholded binary image representing the area covered by the cell
shown in the object 'cellmovie'.}
\format{A 512x512 pixel tiff image. This image contains a single frame.}
\usage{
cellmask
}
\description{
White pixel constitutes a mask representing the area covered
by the cell. Pixel size is 80.0886 nanometer.
}
\keyword{datasets}
