% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellmovie}
\alias{cellmovie}
\title{Sequence of RGB images illustrating the movement of Panx2 foci (in green)
and mitochondria (in red).}
\format{A stack of 512x512 pixel tiff images. This stack contains 3 channels
and 50 frames.}
\usage{
cellmovie
}
\description{
Frames were acquired every 0.186s. Pixel size is 80.0886 nanometer.
}
\keyword{datasets}
