% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloseEnough.R
\docType{package}
\name{CloseEnough}
\alias{CloseEnough}
\alias{CloseEnough-package}
\title{CloseEnough: A package using particle trajectories to compute non-random association
between particles and cellular structures}
\description{
The CloseEnough package provides two main fucntions: randomize_trajectory and min_distance
}
\section{randomize_trajectory}{

The randomize_trajectory functions ...
}

