% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coordinates}
\alias{coordinates}
\title{A data frame containing the trajectory coordinates of the Panx2 foci shown
in the object 'cellmovie'.}
\format{A data frame with 5066 rows and 4 variables.}
\usage{
coordinates

coordinates
}
\description{
A data frame containing the trajectory coordinates of the Panx2 foci shown
in the object 'cellmovie'.

An object of class Traj containing the trajectory parameters of the Panx2 foci shown
in the object 'cellmovie'.
}
\keyword{datasets}
