% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{frame_image}
\alias{frame_image}
\title{Stores the information of an image into a data frame.}
\usage{
frame_image(object)
}
\arguments{
\item{object}{An object of class Image}
}
\description{
Stores the information of an image into a data frame.
}
\examples{

## Load image
data(cellmovie)

## Extract a few frames from the image
df_image <- frame_image(cellmovie)


}
