% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{min_distance}
\alias{min_distance}
\title{Calculates the minimum distance between points of a trajectory and objects from a reference image}
\usage{
min_distance(image, trajectory, channel)
}
\arguments{
\item{image}{An object of class Image}

\item{trajectory}{An object of class Traj}

\item{channel}{A numeric vector indicating which channel to use in the image}
}
\description{
Calculates the minimum distance between points of a trajectory and objects from a reference image
}
