% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{mirror_trajectory}
\alias{mirror_trajectory}
\title{Reflect trajectories}
\usage{
mirror_trajectory(object, axis)
}
\arguments{
\item{object}{An object of class 'Traj'}

\item{axis}{If 'x', trajectories are flipped along the horizontal axis. If 'y' they are flipped along the vertical axis.}
}
\description{
Flip all trajectories 'en bloc' along the vertical or horizontal axis.
}
\examples{
data(cellmask)
data(trajectory)
mirroredtraj <- mirror_trajectory(trajectory, "x")
## Display image (first frame only)
EBImage::display(cellmask, method = "raster")

## Overlay trajectories in black
ntraj = nlevels(trajectory["trajectory"])
for (i in 1:ntraj) {
traj_i <- subset(trajectory["data"], trajectory == i)
lines(traj_i$x, traj_i$y, type="l", col = "black")
}
## Overlay mirrored trajectories in blue
nmirroredtraj = nlevels(mirroredtraj["trajectory"])
for (i in 1:nmirroredtraj) {
traj_i <- subset(mirroredtraj["data"], trajectory == i)
lines(traj_i$x, traj_i$y, type="l", col = "blue")
}

}
