% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mitomask}
\alias{mitomask}
\title{A stack of thresholded binary images representing the area covered by the mitochondria
in the cell shown in the object 'cellmovie'.}
\format{A stack of 512x512 pixel tiff images. This stack contains a single channel
and 50 frames.}
\usage{
mitomask
}
\description{
White pixel constitutes a mask representing the area covered
by the mitochondria in each frame. Pixel size is 80.0886 nanometer.
}
\keyword{datasets}
