% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{number_of_frame}
\alias{number_of_frame}
\title{Returns the number of frames}
\usage{
number_of_frame(object)
}
\arguments{
\item{object}{An object of class Traj}
}
\description{
Returns the number of frames
}
\examples{

data(trajectory)
number_of_frame(trajectory)

}
