% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{number_of_trajectory}
\alias{number_of_trajectory}
\title{Returns the number of trajectories}
\usage{
number_of_trajectory(object)
}
\arguments{
\item{object}{An object of class Traj}
}
\description{
Returns the number of trajectories
}
\examples{

data(trajectory)
number_of_trajectory(trajectory)

}
