% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outline}
\alias{outline}
\title{A data frame containing the x,y coordinates of the outline of the cell shown
in the object 'cellmovie'.}
\format{A data frame with 654 rows and 2 variables.}
\usage{
outline
}
\description{
A data frame containing the x,y coordinates of the outline of the cell shown
in the object 'cellmovie'.
}
\keyword{datasets}
