% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{randomize}
\alias{randomize}
\title{Randomizes trajectories within a delimited area.}
\usage{
randomize(trajectory, mask)

randomize(trajectory, mask)
}
\arguments{
\item{trajectory}{An object of class Traj}

\item{mask}{An object of class Image. The image should be a binary mask.}
}
\description{
This function randomizes the trajectories from an object of class Traj within the white area (intensity = 1)
of a binary image.Trajectories are first randomly rotated around their point of origin and then their origin is
randomly assigned to a xy coordinate with the mask area. Trajectory points that fall out of the mask are then
pruned out.
}
\examples{

## Upload mask and trajectories
data(cellmask)
data(trajectory)

## Randomize trajectories
randtraj <- randomize(trajectory, mask = cellmask)

## Plot trajectories
EBImage::display(cellmask, method = "raster")

## Overlay trajectories in black
ntraj = nlevels(trajectory["trajectory"])
for (i in 1:ntraj) {
traj_i <- subset(trajectory["data"], trajectory == i)
lines(traj_i$x, traj_i$y, type="l", col = "black")
}
## Overlay randomized trajectories in blue
nrandtraj = nlevels(randtraj["trajectory"])
for (i in 1:nrandtraj) {
traj_i <- subset(randtraj["data"], trajectory == i)
lines(traj_i$x, traj_i$y, type="l", col = "blue")
}


}
