% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{rotate_trajectory}
\alias{rotate_trajectory}
\title{Rotate trajectories}
\usage{
rotate_trajectory(object, angle)
}
\arguments{
\item{object}{An object of class 'Traj'}

\item{angle}{A numeric value corresponding to the desired angle of rotation.
If positive, trajectories are rotated clockwise. If negative, they are rotated counterclockwise.}
}
\description{
Rotates all trajectories 'en bloc' using the centre of the image as origin.
}
\examples{
## Upload image and trajectories
data(cellmask)
data(trajectory)

rotatedtraj <- rotate_trajectory(trajectory, 90)
## Display image (first frame only)
EBImage::display(cellmask, method = "raster")

## Overlay trajectories in black
ntraj = nlevels(trajectory["trajectory"])
for (i in 1:ntraj) {
traj_i <- subset(trajectory["data"], trajectory == i)
lines(traj_i$x, traj_i$y, type="l", col = "black")
}
## Overlay rotated trajectories in blue
nrotatedtraj = nlevels(rotatedtraj["trajectory"])
for (i in 1:nrotatedtraj) {
traj_i <- subset(rotatedtraj["data"], trajectory == i)
lines(traj_i$x, traj_i$y, type="l", col = "blue")
}

}
