% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{Traj}
\alias{Traj}
\title{Constructor for the S4 class Traj that represents trajectories.}
\usage{
Traj(trajectory = factor(), frame = integer(), time = numeric(),
  time_unit = character(), x = numeric(), y = numeric(),
  dim_x = integer(), dim_y = integer(), pixel_size = numeric(),
  pixel_unit = character(), ...)
}
\arguments{
\item{trajectory}{A vector of unique trajectory identifiers.}

\item{frame}{A vector of integers corresponding to the image frame of each data point.}

\item{time}{An integer corresponding to time interval between frames.}

\item{time_unit}{Name of the unit of time (e.g "second", "s", "min").}

\item{x}{A numeric vector corresponding to the 'x' coordinates of each data point.}

\item{y}{A numeric vector corresponding to the 'y' coordinates of each data point.}

\item{dim_x}{An integer representing the image width (usually in pixels)}

\item{dim_y}{An integer representing the image height (usually in pixels)}

\item{pixel_size}{An integer representing the pixel size (usually in microns or nanometers)}

\item{pixel_unit}{Character(s) for the unit of pixel (e.g. "nanometer", "nm", "micrometer").}

\item{...}{}
}
\description{
\code{Traj} builds a Traj object.
}
\examples{
data(coordinates)
traj <- Traj(trajectory = coordinates$Trajectory,
frame = coordinates$Frame,
time = 0.186,
time_unit = "s",
x = coordinates$x,
y = coordinates$y,
dim_x = 512,
dim_y = 512,
pixel_size = 80.0886,
pixel_unit = "nanometer")

}
