% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{TrajEucl-class}
\alias{TrajEucl-class}
\alias{.TrajEucl}
\title{S4 TrajEucl class}
\description{
TrajEucl inherits from the S4 class Traj but contains and additional slot representing
the minimal distance between every point of all trajectories and another structures.
}
\section{Slots}{

\describe{
\item{\code{trajectory}}{A vector of factor values representing the trajectory ID.}

\item{\code{frame}}{A numeric vector representing the image frames in which the points
within the trajectories were taken}

\item{\code{time}}{A length-one numeric vector representing the time interval between image frames}

\item{\code{time_unit}}{Character(s) representing the unit of time (e.g. "second", "s", "min)}

\item{\code{x}}{A numeric vector corresponding to the 'x' coordinates of each data point.}

\item{\code{y}}{A numeric vector corresponding to the 'y' coordinates of each data point.}

\item{\code{dim_x}}{An integer representing the image width (usually in pixels)}

\item{\code{dim_y}}{An integer representing the image height (usually in pixels)}

\item{\code{pixel_size}}{An integer representing the pixel size (usually in microns or nanometers)}

\item{\code{pixel_unit}}{Character(s) for the unit of pixel (e.g. "nanometer", "nm", "micrometer").}

\item{\code{min_distance}}{A numeric vector.}
}}

