% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{TrajEucl}
\alias{TrajEucl}
\title{Title}
\usage{
TrajEucl(trajectory = factor(), frame = integer(), time = numeric(),
  time_unit = character(), x = numeric(), y = numeric(),
  min_distance = numeric(), dim_x = integer(), dim_y = integer(),
  pixel_size = numeric(), pixel_unit = character(), ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
